+++ harborapi/client.py
@@ -1,10 +1,13 @@
 import json
 from collections import deque
+from dataclasses import dataclass
 from http.cookiejar import CookieJar
 from pathlib import Path
 from typing import (
     Any,
+    Deque,
     Dict,
+    Iterator,
     List,
     Literal,
     NamedTuple,
@@ -151,10 +154,55 @@ class ResponseLogEntry(NamedTuple):
     """A log entry for an HTTP response."""
 
     url: URL
+    """An httpx.URL object representing the URL of the request. Can be cast to string using `str()`."""
     method: str
+    """The HTTP method used for the request."""
     status_code: int
+    """The HTTP status code of the response."""
     duration: float
+    """The duration of the full request/response cycle in seconds."""
     response_size: int
+    """The size of the response body in bytes."""
+
+    def __repr__(self) -> str:
+        return f"<ResponseLogEntry [{self.method} {self.url} {self.status_code}]>"
+
+
+# NOTE: Could we do the same by subclassing deque?
+# We are re-implementing a lot of sequence methods here.
+@dataclass
+class ResponseLog:
+    """A log of HTTP responses."""
+
+    entries: Deque[ResponseLogEntry]
+
+    def __init__(self, max_logs: Optional[int] = None) -> None:
+        """Initialize the log."""
+        self.entries = deque(maxlen=max_logs)
+
+    def add(self, entry: ResponseLogEntry) -> None:
+        """Add a new entry to the log."""
+        self.entries.append(entry)
+
+    def resize(self, max_logs: int) -> None:
+        """Resize the log to the specified maximum number of entries."""
+        self.entries = deque(self.entries, maxlen=max_logs)
+
+    def clear(self) -> None:
+        """Clear the log."""
+        self.entries.clear()
+
+    def __iter__(self) -> Iterator[ResponseLogEntry]:
+        """Return an iterator over the entries in the log."""
+        return iter(self.entries)
+
+    def __getitem__(self, index: int) -> ResponseLogEntry:
+        """Return the entry at the specified index."""
+        return self.entries[index]
+
+    def __len__(self) -> int:
+        """Return the number of entries in the log."""
+        return len(self.entries)
 
 
 class HarborAsyncClient:
@@ -170,6 +218,7 @@ class HarborAsyncClient:
         validate: bool = True,
         raw: bool = False,
         logging: bool = False,
+        max_logs: Optional[int] = None,
         config: Optional[Any] = None,  # NYI
     ) -> None:
         """Initialize a new HarborAsyncClient with either a username and secret,
@@ -207,6 +256,8 @@ class HarborAsyncClient:
             Takes precedence over `validate` if `raw=True`.
         logging : bool
             Enable client logging with `Loguru`.
+        max_logs : Optional[int]
+            The maximum number of entries to keep in the response log.
         config : Optional[Any]
             (NYI) config
 
@@ -256,7 +307,7 @@ class HarborAsyncClient:
         else:
             logger.disable("harborapi")
 
-        self.response_log = deque()  # type: deque[ResponseLogEntry]
+        self.response_log = ResponseLog(max_logs=max_logs)
 
     def log_response(self, response: Response) -> None:
         """Log the response to a request.
@@ -266,7 +317,7 @@ class HarborAsyncClient:
         response : Response
             The response to log.
         """
-        self.response_log.append(
+        self.response_log.add(
             ResponseLogEntry(
                 url=response.url,
                 method=response.request.method,
@@ -278,7 +329,7 @@ class HarborAsyncClient:
 
     @property
     def last_response(self) -> Optional[ResponseLogEntry]:
-        """Return the last response logged."""
+        """The most recently logged response."""
         try:
             return self.response_log[-1]
         except IndexError: